pyproject.toml
[tool.poetry]
name = "my-blog"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
